+++ Source/Scene/ModelExperimental/ModelExperimentalUtility.js
@@ -30,7 +30,7 @@ ModelExperimentalUtility.getFailedLoadFunction = function (model, type, path) {
     if (defined(error)) {
       message += `\n${error.message}`;
     }
-    model._readyPromise.reject(new RuntimeError(message));
+    return Promise.reject(new RuntimeError(message));
   };
 };